Stop folding item names, restrict text boxes
Apparently, ACE thinks `Laserdesignator` and `laserdesignator` are two
differen items, and would show two items in the Arsenal (but both
leading to the same in-game item). So we stop folding the case of items
from the `assignGear` file. If the MM has tow different case expressions
of the same item, they will dupe, but there is not much we can do.

Further, restrict the text boxes to plaintext as a defensive measure,
and set the font to monospaced, so debug screenshots are more easy to
read.